Api_Escola/requirements.txt
asgiref==3.7.2
astroid==3.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
cPython==0.0.6
dill==0.3.7
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
dnspython==2.4.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.5
mccabe==0.7.0
platformdirs==3.11.0
pylint==3.0.1
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomlkit==0.12.1
typed-ast==1.5.5
urllib3==2.0.6
validate-docbr==1.10.0
wrapt==1.15.0
